requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
plotly==5.15.0
streamlit==1.25.0
tensorflow==2.13.0
requests==2.31.0
folium==0.14.0
seaborn==0.12.2
psycopg2-binary==2.9.7
pymongo==4.4.1
python-dotenv==1.0.0